+++ changelog.md
@@ -17,6 +17,7 @@
 - [#2995](https://github.com/ignite/cli/pull/2995/) Add `ignite network request remove-validator` command.
 - [#2999](https://github.com/ignite/cli/pull/2999/) Add `ignite network request remove-account` command.
 - [#2458](https://github.com/ignite/cli/issues/2458) New `chain serve` command UI.
+- [#2992](https://github.com/ignite/cli/issues/2992) Add `ignite chain debug` command.
 
 ### Changes